+++ packages/api/src/axelar-query/browser.ts
@@ -2,13 +2,19 @@ import { Environment } from "@axelarjs/core";
 
 import ky, { type Options } from "ky";
 
+import { createGMPBrowserClient } from "..";
 import { AxelarQueryAPIClient } from "./isomorphic";
 
 export const createAxelarQueryBrowserClient = (
   env: Environment,
-  options: Options
+  options?: Options
 ) =>
-  AxelarQueryAPIClient.init(env, {
-    target: "browser",
-    instance: ky.extend(options),
-  });
+  AxelarQueryAPIClient.init(
+    {
+      target: "browser",
+      instance: ky.extend(options ?? {}),
+    },
+    {
+      gmpClient: createGMPBrowserClient(env),
+    }
+  );