+++ packages/api/src/axelarscan/axelarscan.spec.ts
@@ -1,16 +1,9 @@
-import { AxelarscanClient, createAxelarscanNodeClient } from "./";
+import { createAxelarscanNodeClient } from "./";
 
 describe("axelarscan", () => {
   describe("searchTransactions", () => {
-    let api: AxelarscanClient;
-
-    beforeEach(() => {
-      api = createAxelarscanNodeClient({
-        prefixUrl: "https://testnet.api.axelarscan.io",
-      });
-    });
-
     test("It should get link transactions", async () => {
+      const api = createAxelarscanNodeClient("testnet");
       const res = await api.getRecentLinkTransactions({ size: 10 });
       expect(res?.length === 10);
     });