+++ packages/api/src/index.ts
@@ -4,5 +4,5 @@ export * from "./axelarscan";
 export * from "./gmp";
 // axelar-query
 export * from "./axelar-query";
-// s3 static configs
-export * from "./s3";
+// axelar-config
+export * from "./axelar-config";