+++ packages/api/src/s3/node.ts
@@ -1,9 +1,0 @@
-import got, { type Options } from "got";
-
-import { S3Client } from "./isomorphic";
-
-export const createS3NodeClient = (options: Partial<Options>) =>
-  S3Client.init({
-    target: "node",
-    instance: got.extend(options),
-  });