+++ packages/transaction-recovery/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @axelarjs/transaction-recovery
 
+## 0.1.0
+
+### Minor Changes
+
+- aac5d21: Refactor api, transaction-recovery & core modules to support isomorphic clients
+
+### Patch Changes
+
+- Updated dependencies [aac5d21]
+  - @axelarjs/core@0.2.0
+  - @axelarjs/api@0.2.0
+
+## 0.0.14
+
+### Patch Changes
+
+- converting gmp and s3 service clients to use dynamic imports as well
+
+## 0.0.13
+
+### Patch Changes
+
+- refactor apiClient dynamic import
+
+## 0.0.12
+
+### Patch Changes
+
+- moving @axelarjs/api in package.json
+
 ## 0.0.11
 
 ### Patch Changes